refactor: features/widgets 레이어 추출 (행동/조합 단위 분리)
- 북마크 기능 → features/bookmark (ui: Add/Edit/List/Saved/Toggle, lib: bookmarkUtils)
- 지역 선택 → features/select-region (RegionSelectButtons + RegionSelectModal)
- 지도 열기 → features/open-map (MapButton)
- HeroBanner → widgets/hero-banner (전용 훅 useCarousel(Progress)도 동반 이동)
- CafesPage(페이지 조합) → widgets/cafe-list
- import 경로 정정, 크로스 슬라이스는 @/ alias
- FSD 방향 위반(entities→features, shared→widgets) 제거

* 잔존: shared/hooks/server 의 카페 데이터 훅이 entities 참조(기존 #88 이슈) — 후속 정리
