+++ src/app/(withNav)/cafes/page.tsx
@@ -1,5 +1,5 @@
 import { getCafeRegions } from '@/entities/cafe/api/api';
-import CafesPage from '@/entities/cafe/ui/cafe/CafesPage';
+import CafesPage from '@/widgets/cafe-list/ui/CafesPage';
 
 export default async function Page() {
   const regions = await getCafeRegions();
