+++ src/app/(withNav)/page.tsx
@@ -1,5 +1,5 @@
 import CafeRecommendation from '@/entities/cafe/ui/CafeRecommendation';
-import HeroBanner from '@/shared/ui/HeroBanner';
+import HeroBanner from '@/widgets/hero-banner/ui/HeroBanner';
 import { pageContainer } from '@/app/(withNav)/Home.css';
 
 export default function Home() {
