+++ src/app/(withoutNav)/cafes/[id]/page.tsx
@@ -2,12 +2,12 @@ import { getCafes } from '@/entities/cafe/api/api';
 import { getCafeDetail } from '@/entities/cafe/api/api';
 import ChevronLeft from '@/shared/assets/Icon/Chevron_Left.svg';
 import BackButton from '@/entities/cafe/ui/BackButton';
-import BookMarkButton from '@/entities/cafe/ui/BookMarkButton';
+import BookMarkButton from '@/features/bookmark/ui/BookMarkButton';
 import FlavorList from '@/entities/cafe/ui/FlavorItem';
 import Footer from '@/entities/cafe/ui/Footer';
 import IconWithHashTag from '@/entities/cafe/ui/IconWithHashTag';
 import MainImageCarousel from '@/entities/cafe/ui/MainImageCarousel';
-import MapButton from '@/entities/cafe/ui/MapButton';
+import MapButton from '@/features/open-map/ui/MapButton';
 import MenuList from '@/entities/cafe/ui/MenuList';
 import OriginList from '@/entities/cafe/ui/OriginList';
 import { RoastingBar } from '@/entities/cafe/ui/RoastingBar';
