+++ src/widgets/hero-banner/lib/useCarousel.ts
