+++ src/widgets/hero-banner/ui/HeroBanner/BannerCarousel.tsx
