+++ src/widgets/hero-banner/ui/HeroBanner/HeroBanner.css.ts
