+++ src/widgets/hero-banner/ui/HeroBanner/index.tsx
